requirements.txt
langchain
langchain-community
langchain_text_splitters
langchain-groq
langchain-huggingface
sentence-transformers
langchain_chroma
python_dotenv